rename test_dists to test_distributions
